comment 수정, 삭제 기능 추가

+++ src/main/java/com/example/travelers/controller/CommentsController.java
@@ -7,11 +7,13 @@ import com.example.travelers.service.BoardsService;
 import com.example.travelers.service.CommentsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
-@RestController
+@Controller
 @RequestMapping("/boards/{boardId}/comments")
 public class CommentsController {
 
@@ -41,10 +43,10 @@ public class CommentsController {
 
     // 특정 게시글의 댓글 조회
     // GET /boards/{boardId}/comments
-    @GetMapping
-    public ResponseEntity<List<CommentsDto>> getCommentsByBoardId(@PathVariable Long boardId) {
-        return ResponseEntity.ok(commentsService.getCommentsByBoardId(boardId));
-    }
+//    @GetMapping
+//    public ResponseEntity<List<CommentsDto>> getCommentsByBoardId(@PathVariable Long boardId) {
+//        return ResponseEntity.ok(commentsService.getCommentsByBoardId(boardId));
+//    }
 
     // 특정 댓글 조회
     // GET /boards/{boardId}/comments/{commentId}
@@ -64,7 +66,18 @@ public class CommentsController {
     // DELETE /boards/{boardId}/comments/{commentId}
     @DeleteMapping("{commentId}")
     public ResponseEntity<MessageResponseDto> deleteComment(@PathVariable Long boardId, @PathVariable Long commentId) {
+        if (!commentsService.isCommentRelatedToBoard(boardId, commentId)) {
+            return ResponseEntity.badRequest().build();
+        }
         MessageResponseDto responseDto = commentsService.deleteComment(boardId, commentId);
         return ResponseEntity.ok(responseDto);
     }
+
+    @GetMapping
+    public String getCommentsPageByBoardId(@PathVariable Long boardId, Model model) {
+        List<CommentsDto> comments = commentsService.getCommentsByBoardId(boardId);
+        model.addAttribute("comments", comments);
+        return "comment"; // HTML 파일의 이름 (확장자 제외)을 반환
+    }
+
 }
